Moving updated metadata generator to generator folder
